[quickview] fixed a bug when restoring windwed position / size.
If you had a maximize window, then close it, reopen it, close it again, and reopen it: at this point, when you un-maximized it, the position and size was the maximized position and size, instead of the previous windowed position and size.

This is now fixed, so windowed state should be reliably preserved in all cases.