[QuickView] another tentative to tackle the trivial (but not so much thanks to Qt being a dick) task of handling the fullscreen and maximized states to correctly restore them.
Although (one of) the point of libraries like Qt is to abstract the various platforms and ensure that the behavior is consistent across platforms, Qt does a really fucking poor job when it comes to QWindow... Events are fired in completly different orders, states are differents, etc.

So basically I had to start handling things differntly depending on the platform. And this is until the next bug, or the next update where they might decide to yet again change how things are done ?

And by the way: using fullscreen for maximized frameless windows on Windows is a fucking pain in the ass. It makes absolutely no sense, and worse of all: it makes the screen flicker and going in/out maximized. My guess is that it's the driver switching between integrated and dedicated GPU, or some other OS stuff when making a window fullscreen, but anyway Qt making the window fullscreen is the culprit here. And of course, there's no mention of this in the doc, we have no clue why they did that.